Add initWebpage module to initialize a game, and trigger the render of its view via pubSubTokens
Currently, this module defines initWebpage() function
that initializes a GameController object and uses
PubSub to subscribe to some pubSubTokens.initGameView
(defined in the pubSubTokens module) to trigger
the render of its game view.

Note that index.js now just calls initWebpage()

On the other hand, the GameController objects now publish a
pubSubTokens.initGameView token once the players are initialized.

Note: initWebpage will be improved later on to implement
a better DOM structure.