Make receiveAttack() method of gameController return an outcome code
Before this, this method returns true or false, in case of hit or miss.

Now, it returns an outcome code, to report the case of sink as well:
0: miss
1: hit
2: hit and sink

The previous tests checking the return values to be truthy or falsy
still pass, since hit and hit & sink include both the hit of a ship.