Implement the playTurn() method of GameController class
This method is private, hence not accessible from the outside.
Therefore, it is not tested via Jest.

The following actions are performed:
1. Get the coordinates to attack: this depends on whether
the current player is human or AI. The case of human
player is not implemented yet, and a sample coordinate
is returned. Instead: this means that the page soon
returns an error (as the human player tries to attack
the same cell twice)

2. Attack the opponent: Some outcome status object is
obtained, containing info about whether there was a hit,
a sink, a game win, etc.

3. Perform actions based on hit or miss outcome. So far,
only a message is displayed in the console. Later on,
this has to update the UI.

4. Check if the player has won and in that case, stop the game.

5. Perform post attack actions, if any. For the current AI player,
this means removing the attacked coordinates from the list of
possible target.

6. Pass the turn to the other player.