Implement human player selection of opponent cell to attack
In GameController class, in #getAttackCoords() method, the
human player part is implemented: this consists of
publishing a pubSubTokensUi.enableAimingOnGameboard(this.#opponent)
token to enable the ui to select the cell to attack.

In PlayerDom class, the subscription to the same
pubSubTokensUi.enableAimingOnGameboard(this.#player)
is done, which triggers the enabling of a click event
listener on the gameboard, to select the cell to attack.
Note that only the current opponent player captures
the token.

The callback attached to that event listener gets the
right cell in which the user has clicked and publish a
pubSubTokens.attackCoordsAcquired token with such
coordinates as arguments, which in turn is captured
by the subscription made in #playTurn() method of
GameController class.