Implement functionality to show attack outcome on opponent GameboardDom
In #attackCoordsAcquiredCallback() method of GameController class,
we first subscribe to pubSubTokens.attackOutcomeShown token,
which will be used to notify when the UI has been updated
to be able to proceed handling the game turn.

Then, the opponent is attacked and through the #showAttackOutcome
method, a pubSubTokensUi.showAttackOutcome(this.#opponent) token
is published to trigger the update of the UI.

So far, a simple cell background color is used to show whether
the attack was a hit or a miss. In the future, icons might be
shown, or even sounds played.