Add ShipDom module to create Ship divs to be placed in the GameboardDom
This module defines the ShipDom class which inherits from
PlaceableObjectDom. A custom style is applied, although
this will be improved upon later.

So far, the function is tested temporarily placing
two ships in the gameboards, at some specified locations,
which are not the actual fleet locations. This last
functionality will be added next.