Use assets as background images for ShipDom instance's divs
For each ship, you should specify in image.js module the files
corresponding to a ship directed toward north and one directed
toward east (as you cannot easily rotate the ShipDom instance's div
background without rotating the div itself). Then, they must be
placed in the ships object, with 'ship_name' as label, and
"N" and "S" as sub-labels, respectively.

Note that the same sample ships from the previous commit are
used, so the actual fleet is not shown with background images yet.