Add showDeployedFleet() method to GameboardDom
This is used to show all the deployed fleet's ships in the
GameboardDom instances' div.

The private helper function #showShip has been written to
create an instance of ShipDom class and place its div
in the board's div.

Temporarily, this function has been called when creating
the GameboardDom object, so all the ships are shown.