Add button to toggle the deployed ship view of current player
A toggle fleet view of current player button is appended to the
header of the GameViewDom objects' div. The attached callback is
binded to 'this' and just publishes a
pubSubTokensUi.toggleDeployedFleetShown(this.#currentplayer) token.

Note that the this.#currentPlayer has been temporarily
hardcoded to player1, but this will be fixed later on.

PlayerDom objects subscribe to the same toggleDeployedFleetShown(player)
token, and call this.#gameboardDiv.obj.toggleDeployedFleet() to toggle
the view of the deploye fleet of the corresponding player.