Add test: an AI player can apply the improvedHuntTarget strategy
The strategy is the same as the huntTarget one. However,
in hunt mode, not all cells are considered, but just the
ones for which:

(row + col) % opponentMinShipSize === 0

opponentMinShipSize is the minimum ship size of the opponent.

Let opponentMinShipSize be 2. The above formula selects the cells
on alternating diagonals. You just need to test these to find
any ship of at least opponentMinShipSize length. The remaining
ship cells are hit through the target mode.

See https://towardsdatascience.com/coding-an-intelligent-battleship-agent-bf0064a4b319