Implement the improvedHuntTarget strategy for the AiPlayer instances
In the first version, in hunt mode, a random cell is selected
among the possibleTargets cells. If the condition is fulfilled
it is returned. Otherwise, a new random one is selected
until a valid one is retrieved.

Also, the opponentMinShipSize is not updated in this first
version, but remains equal to the initial value of 2.