Improve the improvedHuntTarget strategy of AI Players: cimpute the set of cells with correct parity of minimum size
A selectedPossibleTargets Map is kept to contain the cells to be attacked
in hunt mode, with the correct parity. Note that so far this
list is only updated, but not actually used to select the
next cell to attack.

This selectedPossibleTargets Map is initialized at the beginning, and
updated each time a new ship is sunk. Indeed, note that:
- if you sunk a ship, you are necessarily in target mode. If there are
  other possible target cells, you remain in target mode and won't be
  using selectedPossibleTargets.
- otherwise, if there is no other possible target cell, you return to
  hunt mode and will be using the updated selectedPossibleTargets.
- if you are in target mode, and score a hit/miss without sinking a
  ship, you continue not using selectedPossibleTargets.
- if you are in hunt mode, you are using the updated selectedPossibleTargets.

To update the selectedPossibleTargets, consider the fact that if the
minimum map size is minMapSize, you can consider different sets of cells,
given by: (row + col) % minShipSize === offset, for offset from 0 to
minMapSize-1. Then, consider the set with the minimum size. If more than
one set has the same minimum size, choose one randomly.