Add probabilistic and improvedProbabilistic strategy for AI Players
Both strategies make use of a frequency Map which checks which
cells are the most likely to have a ship in it by going through
every possible position for each ship.

The improvedProbabilistic uses a squared frequency Map
for cells belonging to selectedPossibleTargets (similar to improvedHuntTarget)
while still allowing randomness in the selection process.

Also, the tests for AiPlayer have been slightly modified to take
into account the different usage of the randomInt function.