Setup the page to be able to deploy the fleets for each player
The DeployFleetViewDom module is initialized, which will create a div
where the player can choose how to deploy its fleet. Right now, an
empty div is shown, and the constructor just randomly place the fleet's ships.

The initWebpage and GameController modules have been refactored to show
the DeployFleetViewDom's div for the first player, and then for the
second, before beginning a new game. These are triggered via PubSub:
recall, however, that no UI is actually shown in that phase, yet.

IMPORTANT: in GameController.test.js, the test "can deploy the players' fleet"
has been removed. The deployment is now done asynchronously.