Add getNestedElementByClass(point, requiredClass) utility function
A helper function getNestedElementByClass(point, requiredClass) has
been added to get the nested element at a position given by
point = [x, y], which has the css class requiredClass.

This function uses the document.elementsFromPoint method, which returns
an array of all the nested element in a given point.

Then, this function has been used in the event listeners in
GameboardDom module, to get the required element (recall that event
listeners in there are attached to the gameboard div, and exploit
event delegation).