Add startMoveShip() and endMoveShip() methods to Gameboard class to move a deployed ship to a different position
The startMoveShip(name) initializes the move of the ship named "name".
Its direction is saved, and it is reset from the gameboard, becoming
not deployed.

The endMoveShip(name, newSternCoords) finalizes the move of the ship
named "name". Its direction is restored, and it is placed in the
gameboard moving its stern to the cell with coordinates newSternCoords.