Refactor rotateShip method of Gameboard module
Let cellCoords be the array of the coordinates of the cells occupied
by a ship, and relativeToCoord the coordinates of a cell among them.

The two helper methods #offsetRelativeToStern(relativeToCoord,
cellCoords) and #sternFromOffset(relativeToCoord, offset, direction)
are added.

The former computes the offset between the relativeToCoord and the stern
cell coords, while the latter computes the stern coordinates given the
offset.

These two methods are now used in the rotateShip method, and will be
used in a later commit, too.