Refactor Gameboard module to be able to move a deployed ship to a different position relative to a non-stern ship coordinate
In the startMoveShip method, the optional parameter
optionalRelativeToCoords is added. This defines the relativeTo variable,
which defines to which cell the position to move to is referred.

relativeToCoords defaults to the stern coordinates if the
optionalRelativeToCoords is omitted.

Then, an offset from the stern cell is computed, and saved
in #shipOnMoveData.

In the endMoveShip method, the second (optional) parameter is
now called newRelativeCoords: it is still the stern coords if the
offset mentioned above is 0 (ie., if optionalRelativeToCoords is omitted).

Then, the new stern coordinates are computed. Note that if
newRelativeCoords is omitted, we still restore the old stern coordinates.