Disable the pointerdown event listener #startEditingPointerDownCallback while it is handled
The first thing #startEditingPointerDownCallback does after a ship is
being targeted is removing the pointerdown listener itself from the
gameboard div.

When the pointerup event is fired on the gameboard, and its event
listener handled (and the corresponding animations are concluded), the
above mentioned event listener is restored.

This prevents error generating occasionally when clicking on a ship
while the animations of the previous edit operation are not completed yet.