Add a "shadow ship" div to gameboardDom object's div to show the constrained position of the ship div being dragged
A shadowShipDiv is created using the initShadowShipDiv() method in the
ShipDom instance corresponding to the ship being dragged. It is appended
to the dom when the dragging starts, and is removed from the dom when
the dragging ends.

While the dragging is on, a transform translate is applied to move it
to the cell grids corresponding to the current pointed cell (constrained
position).

When visible, the shadowShipDiv color is set to semi-transparent black.
When it is in a forbidden position, its color becomes red. Finally, when
it is out of the gameboard, it gets hidden.