Add messages to GameViewDom view to tell which player starts
The msgP p element in GameViewDom view shows a message with the game
status. Once initialized, it subscribes to some
pubSubTokensUi.setGameStatusMsg (defined in pubSubTokens module) to edit
its text content with the provided message.

In #playGame() method of GameController class, the first player to start
is initialized. Then, a pubSubTokensUi.setGameStatusMsg is published,
with a custom status message to inform about that as argument.

The message depends on whether there is 1 or 2 players, and whether the
first player is the AI or not.