Add messages to GameViewDom view to tell each move's outcome
The msgP p element is updated via PubSub to show the attack outcome
message (either hit, miss, or hit and sunk), as well as the current
player to play. To do that, a pubSubTokensUi.setGameStatusMsg token
is published.

The message depends on whether there is 1 or 2 players, and whether
the first player is the AI or not.

Note, however, that showing a message necessarily slows down the game,
as otherwise it would not be readable. Some additional waiting time
needs to be inserted. Consider adding a button to disable such messages,
and make the gameplay faster.